+++ include/NovelTea/CutsceneSegment.hpp
@@ -21,8 +21,8 @@ public:
 
 	static std::shared_ptr<CutsceneSegment> createSegment(const json &j);
 
-	virtual void setDuration(unsigned int duration);
-	virtual unsigned int getDuration() const;
+	virtual void setDuration(size_t duration);
+	virtual size_t getDuration() const;
 
 	void setScriptOverride(bool override);
 	bool getScriptOverride() const;
@@ -31,7 +31,7 @@ public:
 	const std::string &getScriptOverrideName() const;
 
 private:
-	unsigned int m_duration = 1000;
+	size_t m_duration = 1000;
 	bool m_scriptOverride = false;
 	std::string m_scriptOverrideName;
 };
